backend/requirements.txt
# MediExplain Backend Dependencies
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Env var loading + settings validation
pydantic-settings==2.2.1

# Gemini AI (OCR + text simplification)
google-generativeai==0.7.2

# HTTP client (used by ElevenLabs service)
httpx==0.27.0

# Optional: Cloudinary image hosting
cloudinary==1.40.0

# Optional: multipart form data (FastAPI needs this for file uploads)
python-multipart==0.0.9
